pyproject.toml
[project]
name = "changelog-writer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "langchain>=0.3.27",
    "langchain-gigachat>=0.3.12",
    "langchain-groq>=0.3.8",
    "langchain-ollama>=0.3.10",
    "langchain-openai>=0.3.34",
    "langgraph>=0.6.8",
    "pygithub>=2.8.1",
]
